% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_and_suppl.R
\docType{data}
\name{cervical}
\alias{cervical}
\title{Cervical cancer data}
\format{A data frame with 58 observations and 714 variables (i.e miRNAs of human samples).}
\source{
\url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2880020/#supplementary-material-sec}
}
\description{
Cervical cancer data measures the expressions of 714 miRNAs of human samples. There are 29 tumor and 29 non-tumor cervical
samples and these two groups are treated as two separete classes.
}
\examples{
\dontrun{
data(cervical)
}
}
\references{
Witten, D., et al. (2010) Ultra-high throughput sequencing-based small RNA discovery and discrete statistical biomarker
analysis in a collection of cervical tumours and matched controls. BMC Biology, 8:58
}
