% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot.MLSeq}
\alias{plot,MLSeq,ANY-method}
\title{Plot accuracy results from 'MLSeq' object}
\usage{
\method{plot}{MLSeq}(x, y, ...)

\S4method{plot}{MLSeq,ANY}(x, y, ...)
}
\arguments{
\item{x}{an \code{MLSeq} object returned from \code{classify} function.}

\item{y}{this parameter is not used. Deprecated.}

\item{...}{further arguements. Deprecated.}
}
\description{
This generic function is used to plot accuracy results from 'MLSeq' object returned by
\code{classify} function.
}
\author{
Dincer Goksuluk, Gokmen Zararsiz, Selcuk Korkmaz, Vahap Eldem, Ahmet Ozturk and Ahmet Ergun Karaagaoglu
}
