% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show.MLSeq}
\alias{show,MLSeq-method}
\alias{show,MLSeqModelInfo-method}
\alias{show,MLSeqMetaData-method}
\alias{show,voom.train-method}
\alias{show,discrete.train-method}
\title{Show method for MLSeq objects}
\usage{
show.MLSeq(object)

\S4method{show}{MLSeq}(object)

\S4method{show}{MLSeqModelInfo}(object)

\S4method{show}{MLSeqMetaData}(object)

\S4method{show}{voom.train}(object)

\S4method{show}{discrete.train}(object)
}
\arguments{
\item{object}{an \code{MLSeq} object returned from \code{classify} function.}
}
\description{
Prints out the information from the trained model using \code{classify} function.
}
\seealso{
\code{\link{classify}}
}
