% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods.R
\docType{methods}
\name{trained}
\alias{trained}
\alias{trained,MLSeq-method}
\alias{trained,MLSeqModelInfo-method}
\title{Accessors for the 'trainedModel' slot.}
\usage{
trained(object)

\S4method{trained}{MLSeq}(object)

\S4method{trained}{MLSeqModelInfo}(object)
}
\arguments{
\item{object}{an \code{MLSeq} or \code{MLSeqModelInfo} object.}
}
\description{
This slot stores the trained model. This object is returned from \code{train} function in caret package.
Any further request using caret functions is available for \code{trainedModel} since this object is in the
same class as the returned object from \code{train}. See \code{\link[caret:train]{train}} for details.
}
\examples{
\dontrun{
library(DESeq2)
data(cervical)

# a subset of cervical data with first 150 features.
data <- cervical[c(1:150), ]

# defining sample classes.
class <- data.frame(condition = factor(rep(c("N","T"), c(29, 29))))

n <- ncol(data)  # number of samples
p <- nrow(data)  # number of features

# number of samples for test set (30\% test, 70\% train).
nTest <- ceiling(n*0.3)
ind <- sample(n, nTest, FALSE)

# train set
data.train <- data[ ,-ind]
data.train <- as.matrix(data.train + 1)
classtr <- data.frame(condition = class[-ind, ])

# train set in S4 class
data.trainS4 <- DESeqDataSetFromMatrix(countData = data.train,
                   colData = classtr, formula(~ 1))

## Number of repeats (repeats) might change model accuracies ##
# Classification and Regression Tree (CART) Classification
cart <- classify(data = data.trainS4, method = "rpart",
          ref = "T", preProcessing = "deseq-vst",
          control = trainControl(method = "repeatedcv", number = 5,
                                 repeats = 3, classProbs = TRUE))

trained(cart)
}

}
\seealso{
\code{\link[caret:train]{train.default}}, \code{\link{voom.train-class}}, \code{\link{discrete.train-class}}
}
