% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voomFunctions.R
\name{voomControl}
\alias{voomControl}
\title{Define controlling parameters for voom-based classifiers}
\usage{
voomControl(method = "repeatedcv", number = 5, repeats = 10,
  tuneLength = 10)
}
\arguments{
\item{method}{validation method. Support repeated cross validation only ("repeatedcv").}

\item{number}{a positive integer. Number of folds.}

\item{repeats}{a positive integer. Number of repeats.}

\item{tuneLength}{a positive integer. If there is a tuning parameter in the classifier, this value
is used to define total number of tuning parameter to be searched.}
}
\description{
This function sets the control parameters for voom based classifiers while training the model.
}
\examples{
1L

}
\seealso{
\code{\link{classify}}, \code{\link[caret]{trainControl}}, \code{\link{discreteControl}}
}
\author{
Dincer Goksuluk, Gokmen Zararsiz, Selcuk Korkmaz, Vahap Eldem, Ahmet Ozturk and Ahmet Ergun Karaagaoglu
}
\keyword{RNA-seq}
\keyword{classification}
