\name{cervical}
\alias{cervical}
\docType{data}
\title{
Cervical Cancer Data
}
\description{
Cervical cancer data measures the expressions of 714 miRNAs of human samples. There are 29 tumor and 29 non-tumor cervical samples and these two groups are treated as two separete classes.
}
\usage{data(cervical)}
\format{
  A data frame with 58 observations on the following 715 variables.
}
\source{
http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2880020/#supplementary-material-sec
}
\references{
Witten, D., et al. (2010) Ultra-high throughput sequencing-based small RNA discovery and discrete statistical biomarker analysis in a collection of cervical tumours and matched controls. BMC Biology, 8:58
}
\examples{
data(cervical)
}

