% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_and_suppl.R
\docType{package}
\name{MLSeq-package}
\alias{MLSeq-package}
\title{Machine learning interface for RNA-Seq data}
\description{
This package applies machine learning methods, such as Support Vector Machines (SVM), Random Forest (RF),
Classification and Regression Trees (CART), Linear Discriminant Analysis (LDA) and more to RNA-Seq data. MLSeq combines
well-known differential expression algorithms from bioconductor packages with functions from a famous package \code{caret},
which has comprehensive machine learning algorithms for classification and regression tasks. Although \code{caret} has 200+
classification/regression algorithm built-in, approximately 85 classification algorithms are used in \code{MLSeq} for classifying
gene-expression data. See \code{availableMethods()} for further information.
}
\seealso{
\code{\link{availableMethods}}, \code{\link[caret:modelLookup]{getModelInfo}}

\tabular{ll}{
  Package: \tab MLSeq\cr
  Type: \tab Package\cr
  License: \tab GPL (>= 2)\cr
}
}
\author{
Dincer Goksuluk, Gokmen Zararsiz, Selcuk Korkmaz, Vahap Eldem, Ahmet Ozturk and Ahmet Ergun Karaagaoglu

-----------------

Maintainers:

Dincer Goksuluk \email{dincer.goksuluk@hacettepe.edu.tr}

Gokmen Zararsiz, \email{gokmenzararsiz@erciyes.edu.tr}

Selcuk Korkmaz \email{selcukorkmaz@hotmail.com}
}
\keyword{package}
