\name{MLSeq-package}
\alias{MLSeq-package}
\docType{package}
\title{
Machine Learning Interface for RNA-Seq data
}
\description{
This package applies several machine learning methods, including SVM, bagSVM, Random Forest and CART, to RNA-Seq data. 
}
\details{
\tabular{ll}{
Package: \tab MLSeq\cr
Type: \tab Package\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Gokmen Zararsiz, Dincer Goksuluk, Selcuk Korkmaz, Vahap Eldem, Izzet Parug Duru, Turgay Unver, Ahmet Ozturk

Maintainer: Gokmen Zararsiz \email{gokmenzararsiz@erciyes.edu.tr}
}

