% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/methods.R
\docType{methods}
\name{print}
\alias{print}
\alias{print.confMat}
\alias{print,confMat-method}
\title{print method for confusion matrix}
\usage{
print(x, ...)

\method{print}{confMat}(x, ..., mode = x$mode, digits = max(3,
  getOption("digits") - 3))

\S4method{print}{confMat}(x, ..., mode = x$mode, digits = max(3,
  getOption("digits") - 3))
}
\arguments{
\item{x}{an object of class \code{confMat}}

\item{\dots}{further arguments to be passed to \code{print.table}}

\item{mode}{see \code{\link[caret]{print.confusionMatrix}}}

\item{digits}{see \code{\link[caret]{print.confusionMatrix}}}
}
\description{
This function prints the confusion matrix of the model.
}
